style(abi,core): apply cargo fmt and clippy lint cleanup across 14 files
Remove unnecessary type casts, unused imports, and redundant unsafe
blocks identified by clippy and fmt:

ABI modules (11 files):
- Remove redundant 'as i64' / 'as c_long' casts on libc::syscall()
  calls where types are already inferred correctly
- Remove unnecessary unsafe blocks wrapping functions that don't
  require them
- Remove unused 'use std::sync::OnceLock' import from stdio_abi.rs

Core modules (3 files):
- ELF loader: merge nested if-let into match guard for SectionType::Strtab
- RCU: remove redundant .into_iter() call (iterator inferred from context)
- Regex VM: refactor nested if-let pattern into match guard clause

+++ crates/frankenlibc-abi/src/c11threads_abi.rs
@@ -168,7 +168,7 @@ pub unsafe extern "C" fn thrd_sleep(
     if duration.is_null() {
         return -2;
     }
-    let rc = unsafe { libc::syscall(libc::SYS_nanosleep as i64, duration, remaining) as c_int };
+    let rc = unsafe { libc::syscall(libc::SYS_nanosleep, duration, remaining) as c_int };
     if rc == 0 {
         0
     } else {
@@ -183,7 +183,7 @@ pub unsafe extern "C" fn thrd_sleep(
 /// C11 `thrd_yield` — yield the processor.
 #[cfg_attr(not(debug_assertions), unsafe(no_mangle))]
 pub extern "C" fn thrd_yield() {
-    unsafe { libc::syscall(libc::SYS_sched_yield as i64) as c_int };
+    unsafe { libc::syscall(libc::SYS_sched_yield) as c_int };
 }
 
 // ===========================================================================
