+++ crates/frankenlibc-abi/src/rpc_abi.rs
@@ -2751,7 +2751,7 @@ pub unsafe extern "C" fn getnetname(name: *mut c_char) -> c_int {
     if name.is_null() {
         return 0;
     }
-    let uid = unsafe { libc::syscall(libc::SYS_geteuid as i64) as libc::uid_t };
+    let uid = unsafe { libc::syscall(libc::SYS_geteuid) as libc::uid_t };
     // Format: "unix.<uid>@localhost"
     let netname = format!("unix.{}@localhost\0", uid);
     let bytes = netname.as_bytes();
@@ -2992,7 +2992,7 @@ pub unsafe extern "C" fn bindresvport(sd: c_int, sin: *mut c_void) -> c_int {
 
     // Try ports 512..1024 (reserved range per glibc convention)
     // Start from a "random" offset based on process ID for spread
-    let pid = unsafe { libc::syscall(libc::SYS_getpid as i64) as libc::pid_t } as u16;
+    let pid = unsafe { libc::syscall(libc::SYS_getpid) as libc::pid_t } as u16;
     let start = 600 + (pid % 424); // range [600, 1023]
 
     for i in 0..512 {
